Abstracts WeekInReviewStatistics into own component
